+++ src/routes/graphql/mutations/mutation.types.ts
@@ -1,25 +1,12 @@
 import {
-  GraphQLEnumType,
   GraphQLBoolean,
   GraphQLInt,
   GraphQLInputObjectType,
   GraphQLFloat,
   GraphQLString,
 } from 'graphql';
-import { UUIDType } from './uuid.js';
-import { MemberTypes } from './args.js';
-
-export const MemberTypeId = new GraphQLEnumType({
-  name: 'MemberTypeId',
-  values: {
-    [MemberTypes.Basic]: {
-      value: MemberTypes.Basic,
-    },
-    [MemberTypes.Business]: {
-      value: MemberTypes.Business,
-    },
-  },
-});
+import { UUIDType } from '../types/uuid.js';
+import { MemberTypeId } from '../queries/query.types.js';
 
 export const CreateProfileInput = new GraphQLInputObjectType({
   name: 'CreateProfileInput',
@@ -44,28 +31,27 @@ export const CreateUserInput = new GraphQLInputObjectType({
   name: 'CreateUserInput',
   fields: () => ({
     name: { type: UUIDType },
-    balance: { type: GraphQLFloat }
-  })
+    balance: { type: GraphQLFloat },
+  }),
 });
 
 export const ChangePostInput = new GraphQLInputObjectType({
   name: 'ChangePostInput',
   fields: () => ({
-    title: { type: GraphQLString }
-  })
+    title: { type: GraphQLString },
+  }),
 });
 
 export const ChangeProfileInput = new GraphQLInputObjectType({
   name: 'ChangeProfileInput',
   fields: () => ({
-    isMale: { type: GraphQLBoolean }
-
-  })
+    isMale: { type: GraphQLBoolean },
+  }),
 });
 
 export const ChangeUserInput = new GraphQLInputObjectType({
   name: 'ChangeUserInput',
   fields: () => ({
-    name: { type: GraphQLString }
-  })
+    name: { type: GraphQLString },
+  }),
 });
